feat：改进了代码，独立了command模块
