+++ syncl2r/command/remote.py
@@ -3,7 +3,7 @@ from .app import app
 from ..utils import pprint
 from ..config import get_global_config
 from ..connect_core import Connection
-from syncl2r.sync_core.deploy_core import (
+from ..command_core import (
     stop_last_pids,
     get_remote_log,
     check_still_running,
@@ -21,7 +21,7 @@ def stop_remote():
 
     cfg = get_global_config()
     if cfg.events is not None and cfg.events.stop is not None:
-        conn.exec_cmd_list(cfg.events.stop)
+        conn.cmd.exec_cmd_list(cfg.events.stop)
 
 
 @remote_cmd.command(name="start", help="start remote production process")
@@ -29,7 +29,7 @@ def start_remote():
     cfg = get_global_config()
     if cfg.events is not None and cfg.events.start is not None:
         conn = Connection()
-        conn.exec_cmd_list(cfg.events.start, cfg.file_sync_config.remote_root_path)
+        conn.cmd.exec_cmd_list(cfg.events.start, cfg.file_sync_config.remote_root_path)
 
 
 @remote_cmd.command(name="relaod", help="restart remote task")
