+++ syncl2r/command/test.py
@@ -2,7 +2,7 @@ from .app import app
 from .app import app
 from ..connect_core import Connection
 from ..bash import get_remote_tree
-from ..sync_core.deploy_core import check_still_running
+from ..command_core import check_still_running
 from ..config import get_global_config
 from ..console import pprint
