+++ syncl2r/command_core/__init__.py
@@ -1,0 +1,2 @@
+from .command_core import CommandExector
+from .deploy_core import stop_last_pids, get_remote_log, check_still_running
