+++ syncl2r/command_core/deploy_core.py
